backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["SabriBkg <sabribah@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "6.0.3"
djangorestframework = "3.17.1"
celery = "5.6.3"
redis = "^7.4.0"
django-environ = "^0.13.0"
psycopg2-binary = "^2.9.11"
cryptography = "^46.0.6"
djangorestframework-simplejwt = "^5.5.1"
django-fsm = "^3.0.1"
django-cors-headers = "^4.9.0"
django-redis = "^6.0.0"
drf-spectacular = "^0.29.0"
sentry-sdk = "^2.57.0"


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
